+++ _examples/http-server/custom-listener/unix-reuseport/main_windows.go
@@ -1,7 +1,0 @@
-// +build windows
-
-package main
-
-func main() {
-	panic("windows operating system does not support this feature")
-}