+++ core/netutil/tcp_soreuse_control_wasm.go
@@ -1,0 +1,9 @@
+// +build wasm
+
+package netutil
+
+import "syscall"
+
+func control(network, address string, c syscall.RawConn) error {
+	return nil
+}